refactor: migrate constants (editor) from apps/web to @plane/constants
